Switch to extended PBD and add first constraints
We now apply constraints using  multiple sub-iterations per timestep. The first
constraint is a density constraint which pushes particles away from each other
based on how many are in a neighborhood. This is the first time that we have
movement in the x and z directions.

Known issue: the particles "explode" on the first load, but then fall with less
energy once they are positions are re-intialized using the reset button. The way
in which particles are added to the grid is a potential cause.