[WIP] Switch to extended particle based dynamics
The previous approaches used smoothed particle hydrodynamics directly; this
modeled forces and updated particles based on newton's second law. However, the
approach we want to use is XPBD, where the SPH is part of a constraint function
that we constantly optimize.

However, this refactor has introduced significant bugs. I have fixed many of
them, which were underlying in the previous implementation. Still, there are
current ones that need to be fixed, hence the logging. The current suspect is
the uniform grid, which seems to be storing particles as neighbors that are way
too far away from a queried particle. On the other hand, this may be a symptom,
not a cause of instability.