Fix instability bug
The fix involved flipping the sign of the constraint function gradient because
we impliclty take it with respect to the neighbor particle, not the current
particle. Due to the chain rule, because we redfine the distance vector r to be
normalized(r_j - r_i), we need to propagate the sign.

This commit removes debugging logging and pulls out some constants to the top of
the file.