Add pipe that can be dipped into the fluid
The frontend lets us translate and rotate the pipe, as well as set its
axial rotation in RPM. The frontend sends the pipe position, rotation,
and RPM to the backend for physics calculation. Dragging the pipe into
the fluid produces expected effects.

To-do still is to improve the translation controls of the pipe to be
pipe-aligned rather than world axis aligned, as well as introduce
some visual indication of the RPM.