Initial viscocity, tensile correction, and vorticity
This commit implements sections 4 and 5 of the Macklin and Mueller
paper. However, the features are extremely sensitive to hyperparameter
changes. Currently, we get stable behavior with 5 iterations per loop,
and the hyperparams for velocity and vorticity confinement in the
commit. If one needs to reduce the number of iterations, it will be
necessary to tweak the other hyperparams as well.