First pass at implementing Takahashi et al
Implement viscosity constraints introduced in the paper "Fast Simulation
of Viscous Fluids with Elasticity and Thermal Conductivity Using
Position-Based Dynamics." This commit introduces viscosity and
elasticity constraints which become terms in the global constraint
function.

However, the associated hyperparameters are highly sensitive. Namely,
having a fluid space of size 5.0 rather than 10.0 causes the system
to become unstable. The current selection of hyperparameters results
in low FPS but stable behavior and can be revised in future commits.